util: Add read helpers for u8, u16, u32 and slices
Accessing self.buffer directly can panic if the indexes are out of
bounds. Bounds checking is required for use in BPF, but is also helpful
here to reduce the chance of a panic (also required for BPF).

This commit adds helper functions to access data from the slice that do
implement bounds checks (via slice.get())

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>